review adjustments
